Make OpenGL graphing mode a setting so you can turn it on/off, sped up realtime graphing in Graphing Window.
OpenGL mode is pretty and you'll like it if you have a good video
card. But, don't enable it if you do a lot of complex graphing and
you have a crappy built-in Intel or AMD graphics card - your
performance will suffer. Otherwise, if you have a nice card enable it
for a nice graphical boost and maybe a little more speed.